+++ internal/cryptoutil/cryptoutil.go
@@ -26,10 +26,30 @@ type Attestor interface {
 	Attest() ([]byte, error)
 }
 
+func PublicKey(kms, name string, opts ...pemutil.Options) (crypto.PublicKey, error) {
+	if kms == "" {
+		s, err := pemutil.Read(name, opts...)
+		if err != nil {
+			return nil, err
+		}
+		if pub, ok := s.(crypto.PublicKey); ok {
+			return pub, nil
+		}
+		return nil, fmt.Errorf("file %s does not contain a valid public key", name)
+	}
+
+	k, err := newKMSPublicKey(kms, name)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get public key: %w", err)
+	}
+
+	return k.Public(), nil
+}
+
 // CreateSigner reads a key from a file with a given name or creates a signer
 // with the given kms and name uri.
 func CreateSigner(kms, name string, opts ...pemutil.Options) (crypto.Signer, error) {
-	if kms == "" {
+	if kms == "" || isSoftKMS(kms) {
 		s, err := pemutil.Read(name, opts...)
 		if err != nil {
 			return nil, err
@@ -41,6 +61,10 @@ func CreateSigner(kms, name string, opts ...pemutil.Options) (crypto.Signer, err
 	}
 
 	return newKMSSigner(kms, name)
+}
+
+func isSoftKMS(kms string) bool {
+	return strings.HasPrefix(strings.ToLower(strings.TrimSpace(kms)), "softkms")
 }
 
 // LoadCertificate returns a x509.Certificate from a kms or file
@@ -138,9 +162,13 @@ func IsKMSSigner(signer crypto.Signer) (ok bool) {
 }
 
 // IsX509Signer returns true if the given signer is supported by Go's
-// crypto/x509 package to sign sign X509 certificates. This methods returns true
+// crypto/x509 package to sign X509 certificates. This methods returns true
 // for ECDSA, RSA and Ed25519 keys, but if the kms is `sshagentkms:` it will
 // only return true for Ed25519 keys.
+// TODO(hs): introspect the KMS key to verify that it can actually be
+// used for signing? E.g. for Google Cloud KMS RSA keys can be used for
+// signing or decryption, but only one of those at a time. Trying to use
+// a signing key to decrypt data will result in an error from Cloud KMS.
 func IsX509Signer(signer crypto.Signer) bool {
 	pub := signer.Public()
 	if ks, ok := signer.(*kmsSigner); ok {
@@ -158,6 +186,12 @@ func IsX509Signer(signer crypto.Signer) bool {
 }
 
 type kmsSigner struct {
+	crypto.PublicKey
+	name     string
+	kms, key string
+}
+
+type kmsPublicKey struct {
 	crypto.PublicKey
 	name     string
 	kms, key string
@@ -204,6 +238,44 @@ func newKMSSigner(kms, key string) (*kmsSigner, error) {
 		kms:       kms,
 		key:       key,
 	}, nil
+}
+
+// newKMSPublicKey creates a signer using `step-kms-plugin` as the signer.
+func newKMSPublicKey(kms, key string) (*kmsPublicKey, error) {
+	name, err := plugin.LookPath("kms")
+	if err != nil {
+		return nil, err
+	}
+
+	args := []string{"key"}
+	if kms != "" {
+		args = append(args, "--kms", kms)
+	}
+	args = append(args, key)
+
+	// Get public key
+	cmd := exec.Command(name, args...)
+	out, err := cmd.Output()
+	if err != nil {
+		return nil, exitError(cmd, err)
+	}
+
+	pub, err := pemutil.Parse(out)
+	if err != nil {
+		return nil, err
+	}
+
+	return &kmsPublicKey{
+		PublicKey: pub,
+		name:      name,
+		kms:       kms,
+		key:       key,
+	}, nil
+}
+
+// Public returns the KMS public key
+func (s *kmsPublicKey) Public() crypto.PublicKey {
+	return s.PublicKey
 }
 
 // Public implements crypto.Signer and returns the public key.