+++ CHANGELOG.md
@@ -26,7 +26,53 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ---
 
-## [Unreleased]
+## [v0.25.0] - 2023-09-26
+
+### Added
+
+- Add support for provisioner claim `disableSmallstepExtensions`
+  (smallstep/cli#986)
+- Add support for PowerShell plugins on Windows (smallstep/cli#992)
+- Create API token using team slug (smallstep/cli#980)
+- Detect OIDC tokens issued by Kubernetes (smallstep/cli#953)
+- Add support for Smallstep Managed Endpoint X509 extension
+  (smallstep/cli#989)
+- Support signing a certificate for a private key that can only be used for 
+  encryption with the `--skip-csr-signature` flag in `step certificate create`. 
+  Some KMSs restrict key usage to a single type of cryptographic operation. 
+  This blocks RSA decryption keys from being used to sign a CSR for their public 
+  key. Using the `--skip-csr-signature` flag, the public key is used directly 
+  with a certificate template, removing the need for the CSR signature.
+- Add all AWS identity document certificates (smallstep/certificates#1510)
+- Add SCEP decrypter configuration flags (smallstep/cli#950)
+- Add detection of OIDC tokens issued by Kubernetes (smallstep/cli#953)
+- Add unversioned release artifacts to build (smallstep/cli#965)
+
+### Changed
+
+- Increase PBKDF2 iterations to 600k (smallstep/cli#949)
+- `--kms` flag is no longer used for the CA (signing) key for 
+`step certificate create`. It was replaced by the `--ca-kms` flag 
+(smallstep/cli#942).
+- Hide `step oauth command` on failure (smallstep/cli#993)
+
+### Fixed
+
+- Look for Windows plugins with executable extensions
+  (smallstep/certificates#976)
+- Fix empty ca.json with invalid template data (smallstep/certificates#1501)
+- Fix interactive prompt on docker builds (smallstep/cli#963)
+- `step certificate fingerprint` correctly parse PEM files with non-PEM header
+  (smallstep/crypto#311)
+- `step certificate format` correctly parse PEM files with non-PEM header
+  (smallstep/cli#1006)
+- Fix TOFU flag in `ca provisioner update` (smallstep/cli#941)
+- Make `--team` incompatible with `--fingerprint` and `--ca-url` in 
+  `step ca bootstrap (smallstep/cli#1017)
+
+### Remove
+
+- Remove automatic creation of the step path (smallstep/certificates#991)
 
 ## [v0.24.4] - 2023-05-11